requirements.txt
langchain>=0.2.10
langgraph>=0.2.16
langchain-community>=0.2.10
langchain-google-genai>=1.0.7
google-generativeai>=0.7.2
pydantic>=2.8.2
pandas>=2.2.2
openpyxl>=3.1.5
python-dotenv>=1.0.1
streamlit>=1.36.0
twilio>=9.0.0
apscheduler>=3.10.4
smtplib